Formatted outputs

+++ custom_modules/FileOperator.py
@@ -1,6 +1,7 @@
 import os
 from custom_modules.FileValidator import fileExists, isFile
-from custom_modules.TypeTester import arg_is_a_list, arg_is_none
+from custom_modules.TypeTester import arg_is_a_list as aial, arg_is_none
+from custom_modules.Utils import flatten_dict as fd
 
 
 def delete_file(file_path):
@@ -10,7 +11,7 @@ def delete_file(file_path):
 
 
 def append_file(file_path, list_data):
-    if arg_is_a_list(list_data):
+    if aial(list_data):
         if fileExists(file_path) and isFile(file_path):
             deleted = delete_file(file_path)
 
@@ -43,15 +44,19 @@ def write_dataframe_to_file(file_path, _data):
 
 
 def write_to_file(file_path, _data):
-    if fileExists(file_path):
-        deleted = delete_file(file_path)
+    _string_data = fd(_data)
 
-        if deleted:
+    if not _string_data == None:
+
+        if fileExists(file_path):
+            deleted = delete_file(file_path)
+
+            if deleted:
+                with open(file_path, "w") as f:
+                    f.write(_string_data)
+
+        else:
             with open(file_path, "w") as f:
-                f.write(_data)
+                f.write(_string_data)
 
-    else:
-        with open(file_path, "w") as f:
-            f.write(_data)
-
-    return fileExists(file_path)
+        return fileExists(file_path)
